Squashed commit of the following:
commit 7653a68c4acf5a97490509108cd12ab84429645f
Merge: a589d3f3 de4c0b8e
Author: jo_jo_1206 <jojo1206@protonmail.com>
Date:   Tue Dec 20 18:01:29 2022 +0900

    Merge branch 'Day46_Hong' into Day47_Merge

commit de4c0b8e532293d18d01be7a2544303049798a1d
Author: usernameIcanuse <90302406+usernameIcanuse@users.noreply.github.com>
Date:   Tue Dec 20 18:01:04 2022 +0900

    Update Render_Manager

commit a589d3f338d2adcd2054757f9a48b74291097cd0
Author: jo_jo_1206 <jojo1206@protonmail.com>
Date:   Tue Dec 20 18:00:04 2022 +0900

    머지 완료

commit 55bf10dc20468cd19e944ae88e6ff4df0c5b5565
Merge: 6150d495 a2b09ad4
Author: SongLomin <lominsong@gmail.com>
Date:   Tue Dec 20 17:28:46 2022 +0900

    Merge remote-tracking branch 'origin/Day46_Moon' into Day47_Merge

commit a2b09ad4668021a2a68721bb08773cc7af9d9897
Author: ans75780 <devmgh@naver.com>
Date:   Tue Dec 20 17:27:24 2022 +0900

    Popup Queue 에 개수제한 줬습니다

commit 6150d495f2823e34107f7cbd3b8bf35d469b1faf
Merge: 8c238a5e ead2cb62
Author: SongLomin <lominsong@gmail.com>
Date:   Tue Dec 20 17:26:17 2022 +0900

    Merge remote-tracking branch 'origin/Day46_SOOmIN' into Day47_Merge

commit ead2cb6267d8e6b3d4f77fe385513dacd5f4ecf2
Author: ctq1234567 <103003894+ctq1234567@users.noreply.github.com>
Date:   Tue Dec 20 17:24:36 2022 +0900

    머지를합시다

commit 8c238a5e383b0d3f0db638985ec9698470cea854
Merge: a6dfa2ce 5c822aff
Author: SongLomin <lominsong@gmail.com>
Date:   Tue Dec 20 17:22:06 2022 +0900

    Merge remote-tracking branch 'origin/Day46_An' into Day47_Merge

commit a6dfa2ce05a73cb72f2a1490799351cffd69cc4d
Merge: bfeb9976 cbff5a48
Author: SongLomin <lominsong@gmail.com>
Date:   Tue Dec 20 17:20:38 2022 +0900

    Merge remote-tracking branch 'origin/Day46_Hong' into Day47_Merge

commit bfeb9976f9b3e85aaf5cbd2bcbbbfc3b95057b90
Merge: addb0c26 be3cf1fb
Author: SongLomin <lominsong@gmail.com>
Date:   Tue Dec 20 17:19:19 2022 +0900

    Merge remote-tracking branch 'origin/Day46_조성우' into Day47_Merge

commit 5c822aff32d81ba1057923986ca008ae82ac209e
Author: AnSeungHan <96778286+AnSeungHan@users.noreply.github.com>
Date:   Tue Dec 20 17:18:51 2022 +0900

    Update

commit addb0c26682b78d1102b0df37d3307100e90f9d7
Author: SongLomin <lominsong@gmail.com>
Date:   Tue Dec 20 17:18:29 2022 +0900

    천옷 물리를 위한 NvCloth 관련 코드 추가 중

commit be3cf1fbf4a764158d480bbe31306ad4bc0e13f3
Author: jo_jo_1206 <jojo1206@protonmail.com>
Date:   Tue Dec 20 17:18:11 2022 +0900

    Start Working on Varg Take Exe

commit cbff5a482e6fae26bde64de3cdb7510bf4549f6f
Author: usernameIcanuse <90302406+usernameIcanuse@users.noreply.github.com>
Date:   Tue Dec 20 17:13:27 2022 +0900

    Update Level_GamePlay.cpp

commit 49179c9b03b38a978860d6293dc64dda14064604
Author: usernameIcanuse <90302406+usernameIcanuse@users.noreply.github.com>
Date:   Tue Dec 20 17:12:24 2022 +0900

    Update Light

commit 181647b2d58c5426f0cf6e05a9abd51af26ee45f
Author: jo_jo_1206 <jojo1206@protonmail.com>
Date:   Tue Dec 20 16:27:59 2022 +0900

    clean up other's shit

    added torch effect, edited radial blur

commit 60393d6d48e57215c7cf772267e5c4100396df64
Author: usernameIcanuse <90302406+usernameIcanuse@users.noreply.github.com>
Date:   Tue Dec 20 13:45:21 2022 +0900

    Update Model Texture

commit 777d2ca1d5952fd302c87d7969778e23226dbbc6
Author: jo_jo_1206 <jojo1206@protonmail.com>
Date:   Tue Dec 20 12:54:10 2022 +0900

    Particle does not explode anymore

commit 31e389bfd6290338b44cc5eb3134d92eef3cde37
Author: usernameIcanuse <90302406+usernameIcanuse@users.noreply.github.com>
Date:   Tue Dec 20 03:55:11 2022 +0900

    Update RenderManager

    Add Setting IrradianceMap

commit ec0a36a0e385aa6677ff9e299dca651144d9e60c
Author: usernameIcanuse <90302406+usernameIcanuse@users.noreply.github.com>
Date:   Tue Dec 20 03:43:29 2022 +0900

    Update PBR

    남은 모델들 스페큘러 맵 바인딩 해줘야 함

commit fddeebcc5309134a72fcfdf8e1b102ce6322301f
Author: ans75780 <devmgh@naver.com>
Date:   Tue Dec 20 03:38:43 2022 +0900

    Add to Memory Popup

    사용 시 메모리를 획득하는 팝업 이미지를 생성했습니다. 사용 효과는 추후 추가할 예정입니다.

commit 20736b5fef83195db65880d356ea08c2090ef771
Author: usernameIcanuse <90302406+usernameIcanuse@users.noreply.github.com>
Date:   Tue Dec 20 00:26:03 2022 +0900

    Update Pbr

commit 2549f27694f386ea5b10ea19c157fac313f2e93c
Author: jo_jo_1206 <jojo1206@protonmail.com>
Date:   Mon Dec 19 21:05:14 2022 +0900

    Varg Effect Editted

+++ DX11_Thymesia_Framework/Client/Private/UI_ItemSlot.cpp
@@ -176,7 +176,7 @@ void CUI_ItemSlot::Update_TextInfo()
 {
 	m_tTextInfo.szText = to_wstring(m_pBindedItem.lock()->Get_CurrentQuantity());
 
-	m_tTextInfo.vPosition.x = m_tUIDesc.fX - (max(0, (m_tTextInfo.szText.size() - 1) ) * 10.f);
+	m_tTextInfo.vPosition.x = m_tUIDesc.fX - (max((unsigned long long)0, (m_tTextInfo.szText.size() - 1) ) * 10.f);
 	m_tTextInfo.vPosition.y = m_tUIDesc.fY;
 
 }
